requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
Babel==2.6.0
django-model-utils==3.1.2
django-phonenumber-field==2.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.2
phonenumbers==8.10.10
Pillow==6.0.0
Pygments==2.3.1
six==1.12.0
django-multiselectfield==0.1.8
django-rest-swagger==2.2.0
django-filters
